Major re-factor
This is actually a squash of a couple of experimental branches which
eventually coalesced on this pattern: remove most of the internal
_methods in all the Setting descendents; leave the weak-ref'd Settings
references in place but have Settings do all manipulation of values as
it ultimately needs to track additional files that get pulled in as the
configuration is extracted (and later manipulated).

This also fixed several issues in filtering, and lays the groundwork for
tracing all substantive configuration lines for a given setting (so that
we can determine when usercfg has overridden something we've set).