Refactor exception handlers and main config
Exception handlers are specified to the application so they're now
static methods on the Application class. Meanwhile, Application.config
is now split into config (the distro-specified config) and args (the
user-specified command line configuration). These are also, along with
parser and store, lazily evaluated properties which will serve to
optimize some of the bash completion cases yet to be implemented.