Move config template into its own file
Because comments can be indented in the configuration file, and the boot
configuration and Python's configparser share a common "#" prefix for
comments, any comments in the template were being ignored and not
output. Now config_template refers to the separate file (relative the
main configuration) that contains the configuration template.