Include all parsed lines with conditions
Have the parser record all lines parsed in the configuration, regardless
of whether their conditional filters make them active or not. Also
record which filters apply to each line, and implement set-based
comparisons of conditional filters (e.g. pi3+ is a subset of the pi3
filter). All this is necessary for the next step of permitting
modification of the root config.txt.