Rewrite the output generator
This is a fairly extensive re-write of the output generator to support
contextual generation of output, i.e. that the config generator should
be capable of writing new conditional sections, and the caller should
be able to specify the conditions it wants to apply to any new
configuration values.

It also introduces the notion that the output generator can write to
multiple files (and always to the root configuration as some settings
can only take effect there), along with options to comment and uncomment
lines instead of outright deleting things.