Move version to package root
Importing pkg_resources turns out to be painfully slow; specifically, on
a Pi just running "pibootctl" on its own takes ~1.9 seconds of which ~1
second is just importing pkg_resources, and we only do that to discover
our own version.

In order to maintain the "single source" setup, the version is now in
the root __init__ module, with setuptools' attr: used in setup.cfg to
pull in the version. This doesn't have any dependency issues as an
entry-point is used for main() anyway.