Handle delegation of output more elegantly
Delegation of output from one setting to another had some issues; if the
setting delegating its output was modified but the setting performing
the output was not there were cases where no output was generated. This
commit fixes those errors.

It also prevents repeated commenting out of lines already commented, and
some of the clock limits for the Pi4 GPU which were modified in the
source documentation since last reading.